Decoupled get and create user, closes #381
