+++ utils/users/createUser.ts
@@ -13,11 +13,7 @@ export async function createUser(
   props: CreateUserInput
 ): Promise<DynamoNewUser> {
   const { email, firstName, lastName } = props;
-  const user = await getUserByEmail({ email });
 
-  if (user) {
-    return user;
-  }
   const userId = nanoid(ID_LENGTHS.USER);
   const newUser: DynamoNewUser = {
     PK: `${ENTITY_TYPES.USER}#${userId}`,
